+++ dwds/CHANGELOG.md
@@ -12,6 +12,7 @@
 - Remove test-only code from `sdk_configuration.dart`.
 - Move shared test-only code to a new `test_common` package.
 - Convert unnecessary async code to sync.
+- Allow empty scopes in expression evaluation in a frame.
 
 **Breaking changes**